feat: add react-router and navigation component
